refactor(protocols): move server implementations up
Relocate application-coupled A2A and bridge runtimes into the server tier while preserving legacy imports and pure protocol definitions.
