+++ aragora/protocols/a2a/__init__.py
@@ -1,5 +1,5 @@
 """
-A2A (Agent-to-Agent Protocol) Implementation.
+A2A (Agent-to-Agent Protocol) definitions and client.
 
 Implements the A2A protocol for inter-agent communication,
 allowing Aragora agents to be discovered and invoked by external systems.
@@ -13,6 +13,8 @@ Key concepts:
 - TaskResult: Result of agent work
 """
 
+import importlib
+
 from aragora.protocols.a2a.types import (
     AgentCard,
     AgentCapability,
@@ -23,14 +25,13 @@ from aragora.protocols.a2a.types import (
     ContextItem,
     SecurityCard,
 )
-from aragora.protocols.a2a.server import A2AServer
 
 try:
     from aragora.protocols.a2a.client import A2AClient
 except ImportError:
     # httpx is an optional dependency in lean CI environments. Keep the
     # broader protocols package importable when the client transport is absent.
-    A2AClient = None  # type: ignore[assignment]
+    A2AClient = None  # type: ignore[misc,assignment]
 
 __all__ = [
     # Types
@@ -46,3 +47,12 @@ __all__ = [
     "A2AClient",
     "A2AServer",
 ]
+
+
+def __getattr__(name: str):
+    if name == "A2AServer":
+        runtime = importlib.import_module("aragora.server.a2a_runtime")
+        server = runtime.A2AServer
+        globals()[name] = server
+        return server
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
