+++ scripts/baselines/import_contracts_baseline.json
@@ -65,18 +65,6 @@
         "aragora.pipeline -> aragora.connectors",
         "aragora.pipeline -> aragora.gateway",
         "aragora.pipeline -> aragora.server",
-        "aragora.protocols -> aragora.agents",
-        "aragora.protocols -> aragora.debate",
-        "aragora.protocols -> aragora.events",
-        "aragora.protocols -> aragora.gauntlet",
-        "aragora.protocols -> aragora.knowledge",
-        "aragora.protocols -> aragora.mcp",
-        "aragora.protocols -> aragora.memory",
-        "aragora.protocols -> aragora.observability",
-        "aragora.protocols -> aragora.reasoning",
-        "aragora.protocols -> aragora.resilience",
-        "aragora.protocols -> aragora.server",
-        "aragora.protocols -> aragora.storage",
         "aragora.ranking -> aragora.server",
         "aragora.reasoning -> aragora.server",
         "aragora.security -> aragora.connectors",
